Create folders for react components and CSS styles. Reflect this change in dependent files. Separate NavBar into its own file.
