+++ .github/pull_request_template.md
@@ -1,0 +1,21 @@
+## FEATURE TITLE
+
+Describe your feature here
+
+### Related PRs
+
+Optional - any related PRs you're waiting on, or PRs that will conflict, etc
+
+### Migrations
+
+Optional - if you added anything to the database through migration(s)
+
+### Tests Performed, Edge Cases
+
+Lol we'll see if we test
+
+### Screenshots
+
+Add screenshots!!! If you'd like
+
+CC: @aivantg
