--- src/App.css
@@ -1,41 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  margin-top: 50px;
-  margin-bottom: 20px;
-  height: 40vmin;
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #09d3ac;
-}
-
-/* makeshit nav */
-
-.makeshiftNav li {
-  display: inline;
-  margin: 0px 20px;
-}
-
-.makeshiftNav li a {
-  color: white;
-  text-decoration: none;
-  border: 2px solid white;
-  padding: 10px 20px;
-  border-radius: 15px;
-}
-
-/* =========== */
