+++ src/components/GeneralOwnerDashboard.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom'
+
+export default class GeneralOwnerDashboard extends React.Component {
+	render() {
+		return (
+			<div> 
+				<p>GENERAL OWNER DASHBOARD</p>
+			</div>
+		);
+	}
+}
