+++ src/styles/App.css
@@ -1,0 +1,9 @@
+.App-logo {
+  margin-top: 50px;
+  margin-bottom: 20px;
+  height: 40vmin;
+}
+
+body {
+  background-color: #282c34;
+}
