+++ src/styles/NavBar.css
@@ -1,0 +1,32 @@
+/* Styling for our NavBar component */
+.navBarCont {
+	background: #ffffff;
+	width: 100%;
+	height: 104px;
+}
+
+nav {
+	width: 100%;
+	height: 100%;
+}
+
+nav ul {
+	height: 100%;
+	width: 40%;
+	float: right;
+	list-style: none;
+	margin: 0px 0px;
+	display: flex;
+	align-items: center;
+	justify-content: flex-end;
+}
+
+nav li {
+	margin: 0px 25px;
+}
+
+nav ul a {
+	text-decoration: none;
+	font-weight: 600;
+	color: #000;
+}
